stein c fix gui oder konsole im debug

+++ Solldaten.h
@@ -17,7 +17,7 @@
 
 
 pcl::PointCloud<pcl::PointXYZ>::Ptr sollstein_erstellen(double aufloesung = 0.1, std::string dat="leer");
-std::vector<float> ausgabe_solldaten(std::string dat);
+std::vector<double> ausgabe_solldaten(std::string dat);
 
 typedef struct points_txt_soll
 {
